Fix Warnings Add Time in Step File

+++ code/XFileExporter.cpp
@@ -111,7 +111,7 @@ XFileExporter::XFileExporter(const aiScene* pScene, IOSystem* pIOSystem, const s
 	mSceneOwned = false;
 
 	// set up strings
-	endstr = "\n"; 
+	endstr = "\n";
 
 	// start writing
 	WriteFile();
@@ -509,7 +509,7 @@ std::string XFileExporter::toXFileString(aiString &name)
 {
 	std::string pref = "NN_"; // node name prefix to prevent unexpected start of string
 	std::string str = pref + std::string(name.C_Str());	
-	for (int i=0; i<str.length(); ++i)
+	for (int i=0; i < (int) str.length(); ++i)
 	{
 		if ((str[i] >= 48 && str[i] <= 57) || // 0-9
 			(str[i] >= 65 && str[i] <= 90) || // A-Z
